fix: AsyncStorage 관련 에러 수정
